cli.find_workflows: Also exclude classes without any abstract methods
We might want to have an abstract workflow class that simply overrides
`__init__` or similar. Currently this is impossible as we only filter out
classes determined to be abstract by `inspect.isabstract`, which will not
exclude abstract classes unless they have at least one abstract method. This
commit therefore adds an explicit check for classes with the metaclass
`abc.ABCMeta`, which should be excluded even if `inspect.isabstract` returns
`False`.