Add single quotes around generated cron expressions
The * asterisk is a special character in yaml, and therefore strings using it should be surrounded by quotes, as in the Github Actions documentation. Currently GADK generates cron expressions without quotes, which could possibly lead to errors due to the use of asterisks. This corrects that by generating single quotes around cron expressions.